+++ src/main/java/com/adriano/repositotory/HashtagRepository.java
@@ -8,8 +8,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HashtagRepository extends JpaRepository<Hashtag, Long> {
+public interface HashtagRepository extends JpaRepository<Hashtag, Integer> {
 
-    List<Hashtag> byCategoria(Long categoria);
+    List<Hashtag> byCategoria(int categoria);
     
 }
